Move primary function into sub mod
